requirements.txt
# AWS SDK and Tools
boto3>=1.26.0
awscli>=1.27.0

# Data Processing and Analysis
pandas>=1.5.0
numpy>=1.23.0
scipy>=1.9.0

# Machine Learning
scikit-learn>=1.1.0
joblib>=1.2.0

# API Interaction
requests>=2.28.0
spotipy>=2.22.0

# Data Validation and Type Checking
pydantic>=1.10.0

# Configuration Management
PyYAML>=6.0.0
python-dotenv>=0.21.0

# Testing
pytest>=7.3.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Code Quality
black>=22.10.0
flake8>=5.0.0
pylint>=2.15.0
mypy>=0.990

# Logging and Monitoring
structlog>=22.3.0

# Jupyter (for notebooks)
jupyter>=1.0.0
notebook>=6.5.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# AWS Glue (for local development)
aws-glue-sessions>=1.0.0

# Progress Bars
tqdm>=4.64.0

# Date and Time handling
python-dateutil>=2.8.0

# Documentation
Sphinx>=5.3.0
sphinx-rtd-theme>=1.1.0
